Add a service worker to be installable & offline
The main complexity here is in pre-caching the application with the
worker and making it automatically update, which is accomplished by
including a script that hashes the app files and emits that in such a
way that the service worker includes it.

To simplify making that work, the webapp files are all moved into their
own directory instead of being next to the rust source.