Make the disabled convert button clearer
It's not clear why the "Convert!" button is disabled immediately on page
load, and on slow networks it can take some time for the worker to
become ready. Fix the styling so it actually becomes visually disabled
when disabled, and change the text to provide some improved feedback
when the worker is not ready.